make sure that flexmock_teardown is transparent w.r.t. the underlying teardown
The problem is that flexmock_teardown raises the errors, which interrupts
the teardown process (and, in all likelihood, makes the next tests fail
because of failed cleanup)

We have to run flexmock_teardown *before* the test's own teardown, as
the test's teardown might called some mocked objects, making tests
pass or fail while they should not.

On minitest v5, we simply use the built-in capture_exceptions to register
the exception in the list of failures.

On minitest v4, we catch the exception in before_teardown and raise it
in after_teardown