make all errors under minitest exceptions instead of assertions
Minitest is very aggressive in its backtrace presentation for assertions:
it will only show the line where the assertion was called -- which makes
an awful lot of sense.

In the case of flexmock, this is problematic as the "assertion failure"
(having a method called more than expected for instance), might be
a few steps deeper in the call chain. One therefore needs to have
the whole call stack to relate the test and the error.