distinguish check and assertion failures
The 'check failed' error is used to report problems with trying to
send an invalid message to a mock/spy. It fails right away, and
is basically the equivalent of a no method error.

The 'assertion failed' error is used at verify time.

This distinction is needed for testing frameworks like minitest that
report the two differently. Assertions are reported at the point
of failure (with backtrace filtered), as minitest understands
"assertion failed" as an assertion call that immediately failed
(not the 'check' case, where the assertion and the test are delayed)